test_requirements.txt
Flask==0.10.1
nose==1.3.3
pyOpenSSL==0.14
requests==2.3.0
